split util files into new test files
